omdb_movies_api/omdb_movies/core/custom_permisssions.py
from rest_framework import permissions


class IsOwner(permissions.BasePermission):
    def has_permission(self, request, view):
        if request.method in permissions.SAFE_METHODS:
            return True
        else:
            if request.method == "POST" and request.data:
                return int(request.data["user"]) == request.user.id
            return True

    """
    Object-level permission to only allow updating own favorites
    """
    def has_object_permission(self, request, view, obj):
        if request.method in permissions.SAFE_METHODS:
            return True
        return obj.user == request.user
